Add moving planets
