+++ bgspace.h
@@ -13,3 +13,32 @@ extern STAR stars[STAR_COUNT];
 void init_star();
 void star_update();
 void star_create();
+
+
+
+#define MAX_COMET_COUNT 10
+extern ALLEGRO_BITMAP  *COMET;
+
+typedef struct{
+    float x,y,scale,speed;
+    bool gone;
+}COMETS;
+
+extern COMETS comets[MAX_COMET_COUNT];
+extern int comet_count;
+extern ALLEGRO_BITMAP *PLANET , *PLANETS[5];
+extern bool planet_gone;
+extern ALLEGRO_DISPLAY* disp;
+extern float x1,y_1,x2,y2,x3,y3,x4,y4;
+
+void init_planets();
+void draw_planets();
+void update_planets();
+void init_planet();
+void destroy_planet();
+void update_comet();
+static void create_comet(COMETS *comet);
+static void draw_comet(COMETS *comet);
+void draw_comets();
+static void create_new_comet();
+void trigger_comet();
